+++ src/sdk/resources/index.ts
@@ -2,3 +2,4 @@ export * from './shops';
 export * from './products';
 export * from './invoices';
 export * from './checkout';
+export * from './cryptoWallet';